+++ dev/tests/integration/testsuite/Magento/Catalog/_files/product_with_two_websites_rollback.php
@@ -4,8 +4,6 @@
  * See COPYING.txt for license details.
  */
 
-use Magento\Catalog\Api\Data\ProductInterface;
-
 $objectManager = \Magento\TestFramework\Helper\Bootstrap::getObjectManager();
 
 /** @var \Magento\Framework\Registry $registry */
@@ -15,16 +13,18 @@ $registry->unregister("isSecureArea");
 $registry->register("isSecureArea", true);
 
 /** @var Magento\Store\Model\Website $website */
-$website = $objectManager->create(\Magento\Store\Model\Website::class);
-$website->load('second_website');
-$website->delete();
+$website = $objectManager->get(Magento\Store\Model\Website::class);
+$website->load('second_website', 'code');
+if ($website->getId()) {
+    $website->delete();
+}
 
 /** @var \Magento\Catalog\Api\ProductRepositoryInterface $productRepository */
 $productRepository = $objectManager->create(\Magento\Catalog\Api\ProductRepositoryInterface::class);
 
 try {
     $firstProduct = $productRepository->get('unique-simple-azaza');
-    $firstProduct->delete();
+    $productRepository->delete($firstProduct);
 } catch (\Magento\Framework\Exception\NoSuchEntityException $exception) {
     //Product already removed
 }
